+++ code/project_model/Farming.cs
@@ -12,11 +12,11 @@ namespace project_model
     public class Farming
     {
         private string Name;
-        private int LiquidPrecipitation;
-        private int[] RelativeHumidity;
-        private int WindSpeed;
-        private int MinTemperature;
-        private int MaxTemperature;
+        private double LiquidPrecipitation;
+        private double[] RelativeHumidity;
+        private double WindSpeed;
+        private double MinTemperature;
+        private double MaxTemperature;
 
         public const String path = "..\\..\\..\\..\\code\\Data\\CULTIVOS\\DATOSDECULTIVOS.csv";
         public const String noregister = "S/R";
@@ -24,7 +24,7 @@ namespace project_model
         /**
          * Farming constructor
          */
-        public Farming(String Name, int LiquidPrecipitation, int[] RelativeHumidity, int WindSpeed, int MinTemperature, int MaxTemperature)
+        public Farming(String Name, double LiquidPrecipitation, double[] RelativeHumidity, double WindSpeed, double MinTemperature, double MaxTemperature)
         {
             this.Name = Name;
             this.RelativeHumidity = RelativeHumidity;
@@ -38,10 +38,10 @@ namespace project_model
          * This Gets and Sets important values in this class
          */
         public string GetName { get => Name; set => Name = value; }
-        public int GetLiquidPrecipitation { get => LiquidPrecipitation; set => LiquidPrecipitation = value; }
-        public int[] GetRelativeHumidity { get => RelativeHumidity; set => RelativeHumidity = value; }
-        public int GetWindSpeed { get => WindSpeed; set => WindSpeed = value; }
-        public int GetMinTemperature { get => MinTemperature; set => MinTemperature = value; }
-        public int GetMaxTemperature { get => MaxTemperature; set => MaxTemperature = value; }
+        public double GetLiquidPrecipitation { get => LiquidPrecipitation; set => LiquidPrecipitation = value; }
+        public double[] GetRelativeHumidity { get => RelativeHumidity; set => RelativeHumidity = value; }
+        public double GetWindSpeed { get => WindSpeed; set => WindSpeed = value; }
+        public double GetMinTemperature { get => MinTemperature; set => MinTemperature = value; }
+        public double GetMaxTemperature { get => MaxTemperature; set => MaxTemperature = value; }
     }
 }
